修正了登陆与注册的uid返回类型

+++ internal_proto/microservices/mysqlmanage/internal/logic/userrigsterlogic.go
@@ -59,12 +59,12 @@ func (l *UserRigsterLogic) UserRigster(in *mysqlmanageserver.UserRegisterRequest
 	if err != nil {
 		logx.Error("Rigster rpc error: %v", err)
 		return &mysqlmanageserver.UserRegisterResponse{
-			UserId: int32(uid),
+			UserId: int64(uid),
 		}, err
 	}
 
 	return &mysqlmanageserver.UserRegisterResponse{
-		UserId: int32(uid),
+		UserId: int64(uid),
 	}, nil
 }
